refactor: Modal 컴포넌트와 Detail 컴포넌트 분리
